v1.1

+++ src/App.vue
@@ -1,12 +1,12 @@
 <template>
     <q-layout view="lHh Lpr lFf">
         <q-layout-header>
-            <q-toolbar color="blue-grey">
+            <q-toolbar color="">
                 <q-btn v-if="logged" flat dense round @click="leftDrawerOpen = !leftDrawerOpen" icon="menu"></q-btn>
 
                 <q-toolbar-title>
-                    Moyenne.info
-                    <div slot="subtitle">Version {{ version }}</div>
+                    Moyenne.info <q-chip color="red" dense square v-if="logged && beta">BETA ACCESS</q-chip>
+                    <div slot="subtitle">Version {{ version }} par Lucas Decrock</div>
                 </q-toolbar-title>
 
                 <div v-if="logged">
@@ -20,7 +20,8 @@
                 <img :src="user.photo" style="width: 75px;">
                 <div class="caption q-ml-md">
                     {{ user.prenom+' '+user.nom }}
-                    <br>{{ user.classe.libelle }}
+                    <br>
+                    <q-chip color="primary" dense square v-if="logged && beta">{{ user.classe.libelle }}</q-chip>
                 </div>
             </div>
             <q-list no-border link inset-delimiter>
@@ -29,10 +30,6 @@
                     <q-item to="/" exact>
                         <q-item-side icon="home"/>
                         <q-item-main label="Accueil"/>
-                    </q-item>
-                    <q-item to="/account" exact>
-                        <q-item-side icon="face"/>
-                        <q-item-main label="Compte"/>
                     </q-item>
                     <q-item to="/moyennes" exact>
                         <q-item-side icon="trending_up"/>
@@ -53,6 +50,11 @@
 </template>
 
 <style>
+    .q-toolbar {
+        background: #00c6ff;
+        background: -webkit-linear-gradient(45deg,  #0072ff, #00c6ff);
+        background: linear-gradient(45deg, #0072ff, #00c6ff);
+    }
 </style>
 
 <script>
@@ -83,20 +85,57 @@
         mounted() {
             this.$q.addressbarColor.set('#0069ff');
 
-            if(this.logged) {
-                window.axios.post('E/' + JSON.parse(localStorage.user).id + '/emploidutemps.awp?verbe=get', 'data={"token": "' + localStorage.token + '"}')
+            if (this.logged) {
+                let user = JSON.parse(localStorage.user);
+
+                window.axios.post('E/' + user.id + '/emploidutemps.awp?verbe=get', 'data={"token": "' + localStorage.token + '"}')
                     .then((response) => {
                         if (response.data.code !== 200) {
-                            window.bus.$q.notify({
-                                message: 'Session expirée, merci de vous reconnecter.',
-                                type: 'negative'
-                            });
+                            if (localStorage.credentials && localStorage.user) {
+                                let credentials = JSON.parse(localStorage.credentials);
 
-                            localStorage.clear();
+                                this.$q.loading.show({
+                                    spinner: 'q-spinner-radio',
+                                    message: 'Session expirée, reconnexion en cours ...',
+                                    spinnerColor: 'white'
+                                });
 
-                            window.bus.$emit('logged-out');
+                                window.axios.post('login.awp', 'data={"identifiant": "'+credentials.username+'", "motdepasse": "'+credentials.password+'"}')
+                                    .then((response) => {
+                                        if (response.data.code === 200) {
+                                            localStorage.token = response.data.token;
 
-                            this.$router.push('/login');
+                                            let accounts = response.data.data.accounts[0].profile.eleves;
+
+                                            let account = accounts.find(function(account) {
+                                                return account.id === user.id;
+                                            });
+
+                                            localStorage.user = JSON.stringify(account);
+
+                                            this.$q.loading.hide();
+                                            window.bus.$emit('logged-in');
+
+                                            this.$router.push('/');
+                                        }
+                                    })
+                                    .catch(error => {
+                                        this.$q.notify('Impossible de rafraichir le token.' + error);
+                                        localStorage.clear();
+                                        window.bus.$emit('logged-out');
+                                        this.$q.loading.hide();
+                                        this.$router.push('/login');
+                                    });
+                            } else {
+                                window.bus.$q.notify({
+                                    message: 'Session expirée, merci de vous reconnecter.',
+                                    type: 'negative'
+                                });
+
+                                window.bus.$emit('logged-out');
+
+                                this.$router.push('/login');
+                            }
                         }
                     })
                     .catch(error => {
